{"schema":"decodex/commit/1","summary":"Split autonomy contract modules","authority":"manual"}
